refactor: Victory, WinningResult class winning 패키지로 분류
